inventory_project/inventory_project/urls.py
"""
URL configuration for inventory_project project.
"""

from django.urls import path, include
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from django.conf import settings
from django.conf.urls.static import static

# ✅ DEFINE urlpatterns FIRST
urlpatterns = [
    path("api/", include("warehouses.urls")),
    path("api/auth/login/", TokenObtainPairView.as_view(), name="jwt_login"),
    path("api/auth/refresh/", TokenRefreshView.as_view(), name="jwt_refresh"),
]

# ✅ THEN append static URLs
if settings.DEBUG:
    urlpatterns += static(
        settings.MEDIA_URL,
        document_root=settings.MEDIA_ROOT,
    )
