requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
python-dotenv
Pillow
reportlab
psycopg2-binary
drf-yasg
